[rdk-b][mt7986][wifi-hal][Add tri-band support]
[Description]
Add hal tri-band support, and the function transform radio index to band by checking frequency.
After this patch the hal function should consider tri-band and transform index to band if needed.
And some previous hal may need to be refactored with the change. This will be done after short-term target finished.
WiFi-agent also use hal to determine which index present in which band. WiFi-agent will still work with this change.

[Release-log]
N/A

Change-Id: I4f44f4db61ba1006796834f1223bace56d94cbb3