[rdkb][common][bsp][add rootfs section to secure boot its]
[Description]
add rootfs section to secure boot its. Since the rootfs node is
also signed by mkimage, we need to make sure the node is existed in
its before generating itb. The content inside the node are not signed
by mkimage. Actual rootfs hash are appended after the itb is generated.

[Release-log]
N/A

Change-Id: Id9ecc39ecb4dd1238d8938c00df322d59b8f67e6